Refactor spacy module
